requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
chroma-hnswlib==0.7.6
discord.py==2.6.3
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.4.0
python-dotenv==1.1.1
setuptools==80.9.0
wheel==0.45.1
yarl==1.21.0
langchain-ollama==0.2.0
ollama==0.3.3
langchain==0.3.1
langchain-community==0.3.1
sentence-transformers==3.1.1
